4*4 keyboard and 7-segment display
In this the 4*4 Keyboard is connected to port0 and the 7-segment display is connected to port1. The controls to these 4 7-segment display are connected to port2.
The pressed key is displayed on all 4 7-segment display. To display only on one 7-segment display, changes can be made at controls on port2